renamed packages to lowercase for consistency
