+++ bin/.gitignore
@@ -1,5 +1,8 @@
+/levels/
+/Input/
 /Main.class
+/View/
 /assets/
 /engine/
-/gardenapocalypse/
-/levels/
+/input/
+/view/
