+++ src/assets/Unit.java
@@ -1,4 +1,4 @@
-package Assets;
+package assets;
 
 public interface Unit 
 {
