+++ src/assets/Zombie.java
@@ -1,4 +1,4 @@
-package Assets;
+package assets;
 
 public class Zombie implements Unit{
