+++ src/engine/Combat.java
@@ -1,12 +1,12 @@
-package Engine;
+package engine;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.Map.Entry;
 import java.util.Queue;
 
-import View.*;
 import Assets.*;
+import view.*;
 
 public class Combat 
 {
