+++ src/input/CommandWords.java
@@ -1,4 +1,4 @@
-package Input;
+package input;
 
 import java.util.ArrayList;
 import java.util.Collection;
